♻️ Refactor to hub module
