--- lamin_cli/_rest/README.md
@@ -1,139 +1,0 @@
-# REST CLI
-
-`lamin rest` exposes thin shell wrappers around the LaminHub REST API for the
-currently connected instance.
-
-Use it when you want to inspect an instance, query records, or issue a small
-mutation from a terminal or script without importing `lamindb`.
-
-```bash
-lamin connect laminlabs/lamin-ops
-lamin rest --help
-```
-
-## Metadata
-
-Read instance-level metadata.
-
-```bash
-lamin rest schema
-lamin rest schema --models
-lamin rest schema core artifact
-lamin rest schema core artifact --format json --compact
-lamin rest schema core --include-hidden
-lamin rest schema core artifact --raw --refresh
-
-lamin rest statistics
-lamin rest statistics core
-lamin rest statistics core ulabel
-lamin rest statistics core artifact 123
-```
-
-## Reads
-
-List objects or get one object by numeric id or uid.
-
-```bash
-lamin rest list core artifact --select uid --select key --limit 10
-
-lamin rest list core artifact \
-  --select uid \
-  --select key \
-  --search training \
-  --limit 10
-
-lamin rest get core artifact j2qX8G9a --select uid --select key
-```
-
-Selects can traverse relations.
-
-```bash
-lamin rest list core artifact \
-  --select uid \
-  --select key \
-  --select 'run(transform(uid,key))'
-```
-
-Search fields can also traverse relations. Relation paths use dot notation.
-
-```bash
-lamin rest list core artifact \
-  --search training \
-  --search-in ulabels.name \
-  --limit 10
-```
-
-Filters are JSON objects passed through to the REST API.
-
-```bash
-lamin rest list core record \
-  --filter '{"and":[{"is_type":{"eq":true}},{"name":{"contains":"dataset"}}]}' \
-  --select uid \
-  --select name
-```
-
-## Mutations
-
-Mutations write to the currently connected instance.
-
-```bash
-lamin rest insert core ulabel --objects '{"name":"treated"}'
-
-lamin rest insert core ulabel \
-  --objects '[{"name":"control"},{"name":"treated"}]'
-
-lamin rest upsert core ulabel \
-  --conflict-column name \
-  --objects '[{"name":"treated"}]'
-
-lamin rest update core ulabel abc12345 \
-  --values '{"description":"updated"}'
-
-lamin rest delete core ulabel abc12345
-
-lamin rest delete core ulabel \
-  --objects '[{"name":"control"},{"name":"treated"}]'
-```
-
-Batch updates identify rows by one or more index columns.
-
-```bash
-lamin rest update core project \
-  --index-column uid \
-  --objects '[{"uid":"abc12345","description":"updated"}]'
-```
-
-Batch deletes accept a list of identifier objects.
-
-```bash
-lamin rest delete core recordrecord \
-  --objects '[{"record_id":1,"feature_id":2,"value_id":3}]'
-```
-
-## JSON input
-
-Options that accept JSON also accept `@path` for a file and `-` for stdin.
-
-```bash
-lamin rest insert core project --objects @projects.json
-cat body.json | lamin rest list core artifact --body -
-```
-
-Several options can be repeated or passed as a JSON list.
-
-```bash
-lamin rest list core artifact --select uid --select key
-lamin rest list core artifact --select '["uid","key"]'
-```
-
-## Design
-
-The commands are intentionally low-level wrappers. They build the REST path and
-request body, authenticate through the existing `lamindb-setup` client, and print
-JSON. The REST API remains responsible for model semantics and validation.
-
-`lamin rest schema` is the exception to the always-raw output style: by default it
-prints a compact markdown summary that is easier to use before composing
-`list`/`get` selects. Pass `--format json` for the summarized metadata as JSON or
-`--raw` for the unprocessed schema endpoint response. Schema responses are cached
-by current instance id and schema id; pass `--refresh` to rebuild the local cache.
