+++ tests/rest/test_query.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import json
 
 from click.testing import CliRunner
-from lamin_cli._rest import rest
+from lamin_cli.hub import rest
 
 
 def test_rest_list_constructs_request(monkeypatch):
@@ -13,7 +13,7 @@ def test_rest_list_constructs_request(monkeypatch):
         calls.append((method, path, params, body))
         return [{"uid": "abc123", "key": "sample.parquet"}]
 
-    monkeypatch.setattr("lamin_cli._rest._query.request_json", fake_request_json)
+    monkeypatch.setattr("lamin_cli.hub._query.request_json", fake_request_json)
 
     result = CliRunner().invoke(
         rest,
@@ -69,7 +69,7 @@ def test_rest_get_constructs_request(monkeypatch):
         calls.append((method, path, params, body))
         return {"uid": "abc123", "name": "sample"}
 
-    monkeypatch.setattr("lamin_cli._rest._query.request_json", fake_request_json)
+    monkeypatch.setattr("lamin_cli.hub._query.request_json", fake_request_json)
 
     result = CliRunner().invoke(
         rest,
