+++ tests/rest/test_statistics.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import json
 
 from click.testing import CliRunner
-from lamin_cli._rest import rest
+from lamin_cli.hub import rest
 
 
 def test_rest_statistics_constructs_request(patch_request_json):
